spank/auks: add force_file_ccache option to force legacy credential storage logic
As discussed in https://github.com/hautreux/auks/issues/43, the
credential storage logic introduced in 0.5.0 may cause issues when used
with the file backend configured for ccache in krb5.conf. The issue is
related to the use of an old krb4 filename pattern in libkrb5.

The new option enables to revert back to storing the credential in a
file with '/tmp/krb5cc_%uid%_%jobid%_XXXXXX' as the location pattern.